rename multiple directories and create directories for services
