+++ word-map/src/main/java/com/margot/word_map/controller/rest/AdminController.java
@@ -3,8 +3,8 @@ package com.margot.word_map.controller.rest;
 import com.margot.word_map.dto.AdminDto;
 import com.margot.word_map.dto.RuleDto;
 import com.margot.word_map.dto.response.GetAdminsResponse;
-import com.margot.word_map.service.AdminService;
-import com.margot.word_map.service.RuleService;
+import com.margot.word_map.service.admin.AdminService;
+import com.margot.word_map.service.rule.RuleService;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import lombok.RequiredArgsConstructor;
