+++ word-map/src/main/java/com/margot/word_map/controller/rest/AuthController.java
@@ -5,7 +5,7 @@ import com.margot.word_map.dto.request.ConfirmRequest;
 import com.margot.word_map.dto.response.ConfirmResponse;
 import com.margot.word_map.dto.response.TokenResponse;
 import com.margot.word_map.model.Admin;
-import com.margot.word_map.service.AdminService;
+import com.margot.word_map.service.admin.AdminService;
 import com.margot.word_map.service.auth.AuthService;
 import io.swagger.v3.oas.annotations.ExternalDocumentation;
 import io.swagger.v3.oas.annotations.Operation;
